[EPMDEDP-10122]: fix: Make sure we init "build_number" field for "edp" versioning
This fix requires also change in admin-console.

Background:
Before moving to V1, status field were set by admin-console, which to be honest doesn't make sense.
During migration, we anotaded "status" field as subresource, so it was ignored when admin-console try to create one.
That is a root cause of this bug. As a temporal (probably not) fix, I decide to init this field in operator.

P.S. Run "go mod tidy -compat=1.17", just to cleanup go.sum file.

Change-Id: I7d006ef048fda180bf05bd58c5b4d6893561e7fe