[EPMDEDP-7985]: fix: Do not make empty commits
When building the codebase, we add some configuration to the repository. If the files/directories already existed, we skipped creating the configuration files but made a commit. That's why the commit was empty. Now we check the status of the repository before committing and skip if there is nothing to commit.

Change-Id: Iab48403123cc9ab362b6e63875d70049b0ae97fc