[EPMDEDP-11991]: fix: Import strategy - the first reconciliation fails with a git clone error
In the clone function, a pull command always returns an error, and the codebase reconciliation fails. The second reconciliation succeeds because we already have the cloned repository and do not call the clone function.
We can remove the pull command from the clone function as it is redundant and causes an error.

Change-Id: If75f56d47e43643888b5594577affbb4a04de66b