[EPMDEDP-12096]: refactor: Optimize repositories cloning
For every reconciliation of the CodeBasebranch, we made cloning for CheckCommitHashExists(if the branch is from the hash) and PutBranchInGit. In case we have a lot of CodeBasebranch CRs - it leads to a high load and slow codebase operator start.

Added field git to the CodeBasebranch status. The operator would set it to "branch-created" if the branch was created. The operator skips cloning repositories for the following reconciliation if the field git is set.

This change is backward compatible as git field is optional.

IMPORTANT: Update CodeBasebranch CRD
Change-Id: I2817c3d23cc0672fda8b2aef04fd7dc7ee5fc87b