fix: Set CodebaseBranch failed status (#65)
This change fixes the incorrect status of the branch if
the push to the remote repository fails.
The issue was with the local working git repository.
If the push to the remote repository fails,
the second reconciliation succeeds because the local branch is created,
and the operator doesn't check whether the remote branch exists.
Added a step to clean the git working directory in case the push fails.